pyproject.toml
[project]
name = "job-watch"
version = "0.1.0"
description = "Hosted BA/PM job monitor with Google Sheets tracking."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "gspread>=6.1.2",
  "google-auth>=2.38.0",
  "httpx[http2]>=0.28.1",
  "rich>=13.9.4",
  "typer>=0.15.1,<0.18.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
]

[project.scripts]
job-watch = "job_watch.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/job_watch"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
